TODO: convert 4 vectors (edgesA1, edgesB1, ...) in splitOnCommonEdge to vector<std::reference_wrapper<PhyloTreeEdge>>, and make copies of input trees in getGeodesic

+++ Bipartition.cpp
@@ -30,7 +30,7 @@ void Bipartition::addOne(size_t index) {
     partition[size() - index - 1] = true;
 }
 
-void Bipartition::complement(int numLeaves) {
+void Bipartition::complement(size_t numLeaves) {
     size_t s = size();
     for (size_t i = 0; i < numLeaves; ++i) {
         partition[s - i - 1].flip();
